Http server updated to master

+++ libyui-qt/src/QY2DiskUsageList.cc
@@ -278,7 +278,7 @@ QY2DiskUsageListItem::usedPercent() const
     int percent = 0;
 
     if ( totalSize() != 0 )
-	percent = ( 100 * usedSize() ) / totalSize();
+	percent = int(( 100 * usedSize() ) / totalSize());
 
     return percent;
 }
@@ -379,4 +379,4 @@ QY2DiskUsageListItem::paintPercentageBar( QPainter *		painter,
     }
 }
 
-#include "QY2DiskUsageList.moc"
+
